lint codebase
