+++ .eslintrc.yml
@@ -1,0 +1,2 @@
+extends: airbnb
+parser: babel-eslint
