+++ package.json
@@ -7,6 +7,7 @@
     "@material-ui/core": "^3.5.1",
     "@material-ui/icons": "^3.0.1",
     "lodash": "^4.17.11",
+    "prop-types": "^15.7.2",
     "react": "^16.6.3",
     "react-document-title": "^2.0.3",
     "react-dom": "^16.6.3",
@@ -32,6 +33,13 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^2.0.1"
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^4.3.0",
+    "eslint-plugin-import": "^2.17.3",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-react": "^7.13.0",
+    "gh-pages": "^2.0.1",
+    "prettier": "^1.18.2"
   }
 }
