+++ src/utilities/misc.js
@@ -1,25 +1,27 @@
 /* Get scrollbar width, from https://stackoverflow.com/a/13382873/5161222 */
 export function getScrollbarWidth() {
-    var outer = document.createElement("div");
-    outer.style.visibility = "hidden";
-    outer.style.width = "100px";
-    outer.style.msOverflowStyle = "scrollbar"; // needed for WinJS apps
+  const outer = document.createElement('div');  // eslint-disable-line
+  outer.style.visibility = 'hidden';
+  outer.style.width = '100px';
+  outer.style.msOverflowStyle = 'scrollbar'; // needed for WinJS apps
 
-    document.body.appendChild(outer);
+  // eslint-disable-next-line
+  document.body.appendChild(outer);
 
-    var widthNoScroll = outer.offsetWidth;
-    // force scrollbars
-    outer.style.overflow = "scroll";
+  const widthNoScroll = outer.offsetWidth;
+  // force scrollbars
+  outer.style.overflow = 'scroll';
 
-    // add innerdiv
-    var inner = document.createElement("div");
-    inner.style.width = "100%";
-    outer.appendChild(inner);        
+  // add innerdiv
+  // eslint-disable-next-line
+  const inner = document.createElement('div');
+  inner.style.width = '100%';
+  outer.appendChild(inner);
 
-    var widthWithScroll = inner.offsetWidth;
+  const widthWithScroll = inner.offsetWidth;
 
-    // remove divs
-    outer.parentNode.removeChild(outer);
+  // remove divs
+  outer.parentNode.removeChild(outer);
 
-    return widthNoScroll - widthWithScroll;
+  return widthNoScroll - widthWithScroll;
 }
